Move publicly required types to public models
