+++ packages/x-date-pickers-pro/src/index.ts
@@ -34,6 +34,4 @@ export * from './MobileDateTimeRangePicker';
 // View renderers
 export * from './dateRangeViewRenderers';
 
-export type { UseDateRangeFieldProps } from './internals/models/dateRange';
-
 export * from './models';
