+++ packages/x-date-pickers-pro/src/internals/models/dateTimeRange.ts
@@ -7,12 +7,12 @@ import {
   DateOrTimeViewWithMeridiem,
 } from '@mui/x-date-pickers/internals';
 import { PickerValidDate } from '@mui/x-date-pickers/models';
-import { DayRangeValidationProps } from './dateRange';
 import {
   DateTimeRangeValidationError,
   RangeFieldSection,
   DateRange,
   RangeFieldSeparatorProps,
+  DayRangeValidationProps,
 } from '../../models';
 
 export interface UseDateTimeRangeFieldProps<
