+++ packages/x-date-pickers-pro/src/internals/models/index.ts
@@ -1,4 +1,3 @@
-export * from './dateRange';
 export * from './dateTimeRange';
 export * from './timeRange';
 export * from './rangePickerProps';
